+++ src/InfluxDB.LineProtocol/Client/LineProtocolClient.cs
@@ -1,8 +1,9 @@
 ﻿using InfluxDB.LineProtocol.Payload;
 using System;
 using System.IO;
-using System.Net;
+using System.IO.Compression;
 using System.Net.Http;
+using System.Net.Http.Headers;
 using System.Net.Sockets;
 using System.Text;
 using System.Threading;
@@ -13,9 +14,10 @@ namespace InfluxDB.LineProtocol.Client
     public class LineProtocolClient : LineProtocolClientBase
     {
         private readonly HttpClient _httpClient;
+        private readonly bool _enableCompression;
 
-        public LineProtocolClient(Uri serverBaseAddress, string database, string username = null, string password = null)
-            : this(new HttpClientHandler(), serverBaseAddress, database, username, password)
+        public LineProtocolClient(Uri serverBaseAddress, string database, string username = null, string password = null, bool enableCompression = false)
+            : this(new HttpClientHandler(), serverBaseAddress, database, username, password, enableCompression)
         {
         }
 
@@ -24,7 +26,8 @@ namespace InfluxDB.LineProtocol.Client
                 Uri serverBaseAddress,
                 string database,
                 string username,
-                string password)
+                string password,
+                bool enableCompression)
             :base(serverBaseAddress, database, username, password)
         {
             if (serverBaseAddress == null)
@@ -34,6 +37,7 @@ namespace InfluxDB.LineProtocol.Client
 
             // Overload that allows injecting handler is protected to avoid HttpMessageHandler being part of our public api which would force clients to reference System.Net.Http when using the lib.
             _httpClient = new HttpClient(handler) { BaseAddress = serverBaseAddress };
+            _enableCompression = enableCompression;
         }
 
         protected override async Task<LineProtocolWriteResult> OnSendAsync(
@@ -64,7 +68,21 @@ namespace InfluxDB.LineProtocol.Client
                     break;
             }
 
-            var content = new StringContent(payload, Encoding.UTF8);
+            HttpContent content;
+
+            if (_enableCompression)
+            {
+                var compressed = Compress(Encoding.UTF8.GetBytes(payload));
+                
+                content = new ByteArrayContent(compressed);
+                content.Headers.ContentEncoding.Add("gzip");
+                content.Headers.ContentType = new MediaTypeHeaderValue("text/plain") { CharSet = "utf-8" };
+            }
+            else
+            {
+                content = new StringContent(payload, Encoding.UTF8);
+            }
+
             var response = await _httpClient.PostAsync(endpoint, content, cancellationToken).ConfigureAwait(false);
             if (response.IsSuccessStatusCode)
             {
@@ -79,6 +97,17 @@ namespace InfluxDB.LineProtocol.Client
             }
 
             return new LineProtocolWriteResult(false, $"{response.StatusCode} {response.ReasonPhrase} {body}");
+        }
+
+        private byte[] Compress(byte[] input)
+        {
+            using (var ms = new MemoryStream())
+            {
+                using (var gz = new GZipStream(ms, CompressionLevel.Fastest))
+                    gz.Write(input, 0, input.Length);
+
+                return ms.ToArray();
+            }
         }
     }
 }